Extract Calculations from Fraction
